infra: Move Polygonmap.js from src/Polygonmap to src
