+++ package.json
@@ -7,7 +7,7 @@
     "maps",
     "polygonmap"
   ],
-  "main": "src/Polygonmap/Polygonmap.js",
+  "main": "src/Polygonmap.js",
   "scripts": {
     "dev": "webpack-dev-server --mode development --config webpack.development.js --open",
     "build": "npm run build-module && npm run build-examples",
