+++ src/utils/colorize/colorize.test.js
