+++ src/utils/normalizeFeature.js
