+++ src/utils/normalizeFeature.test.js
